// frontend/src/types/generated/AttendeeEntryPacket.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AttendeeEntryPacket = { request_id: string, room_id: string, room_uid: string, invite_id: string, expires_at: string, display_name: string, provider: string, attend_command: string, join_url: string, };